Finish renaming cli-related functionality
