+++ packages/cli/src/CliUtils.ts
@@ -2,8 +2,8 @@ import fs from "fs";
 import type { AndroidDetails, App, Publisher, Release, ReleaseJsonMetadata } from "@solana-mobile/dapp-store-publishing-tools";
 import type { Connection } from "@solana/web3.js";
 import { Keypair, PublicKey } from "@solana/web3.js";
-import type { CLIConfig } from "./config/index.js";
-import { getConfig } from "./config/index.js";
+import type { CLIConfig } from "./config/CliConfig.js";
+import { getConfig } from "./config/CliConfig.js";
 import debugModule from "debug";
 import { dump } from "js-yaml";
 import * as util from "util";
