+++ packages/cli/src/commands/create/CreateCliPublisher.ts
@@ -10,7 +10,7 @@ import {
   getConfigWithChecks,
   getMetaplexInstance,
   saveToConfig,
-} from "../../utils.js";
+} from "../../CliUtils.js";
 
 const createPublisherNft = async (
   {
