+++ packages/cli/src/commands/index.ts
@@ -1,3 +1,3 @@
 export * from "./create/index.js";
 export * from "./publish/index.js";
-export * from "./validate.js";
+export * from "./ValidateCommand.js";
