+++ packages/cli/src/commands/scaffolding/index.ts
@@ -1,1 +1,1 @@
-export * from "./init.js";
+export * from "./ScaffoldInit.js";
